REFRACT: - Unificando controladores
